# ATA driver: improved performance