Hi @ghaerr 

I disassembled the code in the ATA driver and found that for the 16-bit code, GCC was doing a lookup for each buffer[i] mention (hence, 2 in the 16-bit code).

So, I've switched to using an incremental pointer. The old code was 10 instructions, while this improved version is 5 instructions. The performance improvements are as follows:

## 8-bit original
106s (format FAT32; 4GB)
24s   (write 2Mb file)
11s    (read 2Mb file)

## 8-bit improved
102s
23s
10s

## 16-bit original
106s
24s
11s

## 16-bit improved
92s (thats 14 seconds saved)
22s
10s

I suspect the changes would be even more obvious if I had tested even larger files.