@fhendrikx 2025-07-14
> Wow, nice work finding and measuring this large bottleneck! I've been disassembling various routines as well, and with all the heavy optimizations already being performed by GCC within this driver I'm surprised it didn't automatically handle this one.

Me too.
 
> I had been thinking of using REP INS/OUT for fast I/O here - given your findings, I'd say it'd be worth trying for even further improvement. This can probably be done pretty easily in C with the asm extension, but given the new code which uses a far buffer, DS will have to be saved and reloaded first, somewhat complicating the asm code.

Yep, agreed. I thought I would try the simple optimisations first.

> I have kept #2352 outstanding awaiting XTIDE test results, I'd like to commit it first before this PR. That will likely cause merge conflicts since the port I/O has been moved into generalized routines, so you may have to fix and resubmit. I'll commit #2352 today so you can more easily move this forward.

Of course. Will adjust it as the changes come through.