@ghaerr 2025-07-18
Hi @fhendrikx,

I ended up having to rewrite the read/write I/O based on newly-learned XTIDE information in PR #2359, and that involved getting right in the middle of this PR's changes. As a result, I ended up adding your enhanced performance code to that PR as well. As such, I'm closing this now.

Thank you!